[FEAT/#174] 에러메시지 분기 처리 및 실패 다이얼로그 상태관리 추가

+++ app/src/main/java/com/sopt/clody/presentation/ui/diarylist/screen/DiaryListViewModel.kt
@@ -5,6 +5,10 @@ import androidx.lifecycle.viewModelScope
 import com.sopt.clody.data.repository.DailyDiaryListRepository
 import com.sopt.clody.data.repository.DiaryListRepository
 import com.sopt.clody.presentation.utils.extension.getDayOfWeek
+import com.sopt.clody.presentation.utils.network.ErrorMessages.FAILURE_NETWORK_MESSAGE
+import com.sopt.clody.presentation.utils.network.ErrorMessages.FAILURE_TEMPORARY_MESSAGE
+import com.sopt.clody.presentation.utils.network.ErrorMessages.UNKNOWN_ERROR
+import com.sopt.clody.presentation.utils.network.NetworkUtil
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -14,7 +18,8 @@ import javax.inject.Inject
 @HiltViewModel
 class DiaryListViewModel @Inject constructor(
     private val diaryListRepository: DiaryListRepository,
-    private val dailyDiaryListRepository: DailyDiaryListRepository
+    private val dailyDiaryListRepository: DailyDiaryListRepository,
+    private val networkUtil: NetworkUtil
 ) : ViewModel() {
 
     private val _diaryListState = MutableStateFlow<DiaryListState>(DiaryListState.Idle)
@@ -24,15 +29,42 @@ class DiaryListViewModel @Inject constructor(
     val selectedDiaryDate: StateFlow<DiaryDate> = _selectedDiaryDate
 
     private val _diaryDeleteState = MutableStateFlow<DiaryDeleteState>(DiaryDeleteState.Idle)
-    val diaryDeleteState: StateFlow<DiaryDeleteState> get() = _diaryDeleteState
+    val diaryDeleteState: StateFlow<DiaryDeleteState>  = _diaryDeleteState
+
+    private val _showDiaryDeleteFailureDialog = MutableStateFlow(false)
+    val showDiaryDeleteFailureDialog: StateFlow<Boolean> = _showDiaryDeleteFailureDialog
+
+    private val _failureDialogMessage = MutableStateFlow("")
+    val failureDialogMessage: StateFlow<String> = _failureDialogMessage
+
+    private val maxRetryCount = 3
+    private var retryCount = 0
 
     fun fetchMonthlyDiary(year: Int, month: Int) {
-        _diaryListState.value = DiaryListState.Loading
+        if (retryCount >= maxRetryCount) return
         viewModelScope.launch {
+            if (!networkUtil.isNetworkAvailable()) {
+                _diaryListState.value = DiaryListState.Failure(FAILURE_NETWORK_MESSAGE)
+                return@launch
+            }
+            _diaryListState.value = DiaryListState.Loading
             val result = diaryListRepository.getMonthlyDiary(year, month)
             _diaryListState.value = result.fold(
-                onSuccess = { DiaryListState.Success(it) },
-                onFailure = { DiaryListState.Failure(it.message ?: "Unknown error") }
+                onSuccess = {
+                    retryCount = 0
+                    DiaryListState.Success(it)
+                },
+                onFailure = {
+                    retryCount++
+                    if (retryCount >= maxRetryCount) {
+                        DiaryListState.Failure(FAILURE_TEMPORARY_MESSAGE)
+                    } else {
+                        val errorMessage = if(it.message?.contains("200") == false) {
+                            FAILURE_TEMPORARY_MESSAGE
+                        } else { it.localizedMessage ?: UNKNOWN_ERROR }
+                        DiaryListState.Failure(errorMessage)
+                    }
+                }
             )
         }
     }
@@ -48,6 +80,12 @@ class DiaryListViewModel @Inject constructor(
 
     fun deleteDailyDiary(year: Int, month: Int, day: Int) {
         viewModelScope.launch {
+            if(!networkUtil.isNetworkAvailable()) {
+                _failureDialogMessage.value = FAILURE_NETWORK_MESSAGE
+                DiaryDeleteState.Failure(_failureDialogMessage.value)
+                _showDiaryDeleteFailureDialog.value = true
+                return@launch
+            }
             _diaryDeleteState.value = DiaryDeleteState.Loading
             val result = dailyDiaryListRepository.deleteDailyDiary(year, month, day)
             _diaryDeleteState.value = result.fold(
@@ -55,10 +93,21 @@ class DiaryListViewModel @Inject constructor(
                     DiaryDeleteState.Success
                 },
                 onFailure = {
-                    DiaryDeleteState.Failure(it.message ?: "Unknown error")
+                    _failureDialogMessage.value = if (it.message?.contains("200") == false) {
+                        FAILURE_TEMPORARY_MESSAGE
+                    } else {
+                        it.localizedMessage ?: UNKNOWN_ERROR
+                    }
+                    _showDiaryDeleteFailureDialog.value = true
+                    DiaryDeleteState.Failure(_failureDialogMessage.value)
                 }
             )
         }
+    }
+
+    fun dismissDiaryDeleteFailureDialog() {
+        _showDiaryDeleteFailureDialog.value = false
+        _failureDialogMessage.value = ""
     }
 
     data class DiaryDate(
